requirements.txt
boto3==1.28.0
Pillow==9.5.0
python-dotenv==1.0.0
reportlab==3.6.12